refactor: rename command files
