+++ src/main/java/ass/nerdy/autosniper/modules/commands/ModeCmd.java
@@ -12,49 +12,52 @@ import java.io.File;
 import java.io.FileReader;
 import java.io.FileWriter;
 import java.util.HashMap;
+import java.util.Map;
 
-public class fS extends CommandBase {
+public class ModeCmd extends CommandBase {
     private final pC checker;
+    private final Map<String, String> modeMap = new HashMap<>();
     private static final File CONFIG_FILE = new File("config/autosniper.json");
     private static final Gson GSON = new GsonBuilder().setPrettyPrinting().create();
-    private final String prefix = EnumChatFormatting.GRAY + "[" + EnumChatFormatting.LIGHT_PURPLE + "N" + EnumChatFormatting.GRAY + "] ";
+    final String prefix = EnumChatFormatting.GRAY + "[" + EnumChatFormatting.LIGHT_PURPLE + "N" + EnumChatFormatting.GRAY + "] ";
 
-    public fS(pC checker) {
+    public ModeCmd(pC checker) {
         this.checker = checker;
+        modeMap.put("1s", "/play bedwars_eight_one");
+        modeMap.put("2s", "/play bedwars_eight_two");
+        modeMap.put("3s", "/play bedwars_four_three");
+        modeMap.put("4s", "/play bedwars_four_four");
+        modeMap.put("4v4", "/play bedwars_two_four");
     }
 
     @Override
     public String getCommandName() {
-        return "minfkdr";
+        return "mode";
     }
 
     @Override
     public String getCommandUsage(ICommandSender sender) {
-        return "/minfkdr [fkdr value]";
+        return "/mode [mode]";
     }
 
     @Override
     public void processCommand(ICommandSender sender, String[] args) {
         if (args.length == 0) {
-            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.RED + "You must provide an FKDR value to set!"));
+            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.RED + "You must provide a mode to set!"));
             return;
         }
 
-        try {
-            double fkdrValue = Double.parseDouble(args[0]);
+        String input = args[0].toLowerCase();
+        String command = modeMap.getOrDefault(input, String.join(" ", args));
 
-            JsonFKDR(fkdrValue);
+        checker.setAutoRqCommand(command);
 
-            checker.setFkdrValue(String.valueOf(fkdrValue));
+        JsonAutoRq(command);
 
-            sender.addChatMessage(
-                    new ChatComponentText(prefix + EnumChatFormatting.GREEN + "Minimum FKDR set to: " + EnumChatFormatting.AQUA + fkdrValue));
-        } catch (NumberFormatException e) {
-            sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.RED + "Invalid number. Please enter a valid FKDR."));
-        }
+        sender.addChatMessage(new ChatComponentText(prefix + EnumChatFormatting.GREEN + "Auto-RQ mode set to: " + EnumChatFormatting.AQUA + command));
     }
 
-    private void JsonFKDR(double fkdrValue) {
+    private void JsonAutoRq(String command) {
         try {
             if (!CONFIG_FILE.getParentFile().exists()) {
                 CONFIG_FILE.getParentFile().mkdirs();
@@ -67,7 +70,7 @@ public class fS extends CommandBase {
                 }
             }
 
-            config.put("fkdrValue", fkdrValue);
+            config.put("autoRqCommand", command);
 
             try (FileWriter writer = new FileWriter(CONFIG_FILE)) {
                 GSON.toJson(config, writer);
