+++ src/main/java/ass/nerdy/autosniper/modules/commands/TargetCmd.java
@@ -9,7 +9,7 @@ import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 
-public class bL extends CommandBase {
+public class TargetCmd extends CommandBase {
     private final String prefix = EnumChatFormatting.GRAY + "[" + EnumChatFormatting.LIGHT_PURPLE + "N" + EnumChatFormatting.GRAY + "] ";
 
     @Override
